Move admin folders into components folder
